Img32.CQ   significantly rewrite   now defaults to MedianCut algorithm for color reduction Img32.Fmt.PNG   fixed several minor image loading bugs Img32.Transform   fixed a minor bug in TWeightedColor

+++ source/Img32.Transform.pas
@@ -864,7 +864,7 @@ end;
 
 procedure TWeightedColor.Add(c: TColor32; w: Integer);
 var
-  a: Integer;
+  a: Int64;
   argb: TARGB absolute c;
 begin
   inc(fAddCount, w);
@@ -889,7 +889,7 @@ end;
 
 procedure TWeightedColor.Subtract(c: TColor32; w: Integer);
 var
-  a: Integer;
+  a: Int64;
   argb: TARGB absolute c;
 begin
   dec(fAddCount, w);
